+++ sandbox/pages/SandboxDropdown.vue
@@ -164,6 +164,7 @@
           </KButton>
           <template #items>
             <KDropdownItem
+              data-testid="router-link"
               :item="{ label: 'Home', to: { name: 'home' } }"
               target="_blank"
             >
@@ -171,6 +172,7 @@
               Router link
             </KDropdownItem>
             <KDropdownItem
+              data-testid="external-link"
               :item="{ label: 'External link', to: 'https://kongponents.konghq.com/' }"
               target="_blank"
             >
@@ -178,6 +180,7 @@
               External link
             </KDropdownItem>
             <KDropdownItem
+              data-testid="disabled-button"
               disabled
               has-divider
               @click="handleItemClick"
@@ -186,6 +189,7 @@
               Disabled button
             </KDropdownItem>
             <KDropdownItem
+              data-testid="disabled-router-link"
               disabled
               :item="{ label: 'Home', to: { name: 'home' } }"
             >
@@ -193,6 +197,7 @@
               Disabled router link
             </KDropdownItem>
             <KDropdownItem
+              data-testid="disabled-external-link"
               disabled
               :item="{ label: 'External link', to: 'https://kongponents.konghq.com/' }"
             >
@@ -201,6 +206,7 @@
             </KDropdownItem>
             <KDropdownItem
               danger
+              data-testid="button"
               has-divider
               @click="handleItemClick"
             >